#19 - fixed bug on form when not declare initial valur for form and add validation
